requirements.txt
numpy  < 1.22
matplotlib>=3.3.4
netcdf4>=1.5.6
setuptools>=52.0.0
psutil>=5.8.0
numba >= 0.5
pyproj>=3.0
scipy>=1.6.1
pyyaml >= 5.3.1
python-dateutil >= 2.8